requirements.txt
pre-commit==3.7.0
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
ruff==0.3.4
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-utils==0.41.2
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
python-multipart==0.0.9
alembic==1.13.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-sugar==1.0.0
httpx==0.27.0
factory_boy==3.3.0
